requirements.txt
Django==3.2.10
axblack==20210720
djangorestframework==3.12
markdown==3.0
django-filter==1.0.1
django-cors-headers==3.10.1
django-jsonfield==1.4.1
django-utils-six==2.0
pusher==3.1.0
ptvsd==4.3.2
